+++ app/assets/config/manifest.js
@@ -1,5 +1,4 @@
 //= link_tree ../images
-//= link_directory ../javascripts .js
 //= link_tree ../builds
 //= link_tree ../../javascript .js
 //= link_tree ../../../vendor/javascript .js
