+++ app/javascript/channels/consumer.js
@@ -1,0 +1,3 @@
+import { createConsumer } from "@rails/actioncable"
+
+export default createConsumer()
