+++ app/views/projects/index.html.haml
@@ -1,6 +1,3 @@
-- content_for :head do
-  = javascript_include_tag 'application'
-
 #wrapper.projects
   %h3 Today is a great day to make some updates!
